requirements.txt
pyyaml
numpy~=1.24.3
scipy~=1.10.1
polars~=0.18.7
pandas[performance]~=2.0.2
kaggle~=1.5.13
pyarrow~=12.0.0
hydra-core~=1.3.0
dvc[gdrive]~=3.0
cupy-cuda11x~=12.1.0
seaborn~=0.12.2
scikit-learn~=1.2.2
pynndescent~=0.5.10
hydra-optuna-sweeper~=1.2.0
